# web/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Rempart DNS",
    "version": "1.1",
    "description": "API REST de Rempart. Session (cookie rempart_session + en-tête X-Rempart: 1 sur les requêtes qui modifient) ou jeton porteur rmp_…. Les champs inconnus sont refusés (400)."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/groups": {
      "get": {
        "tags": [
          "Groupes d'appareils"
        ],
        "summary": "Groupes, services et catégories disponibles",
        "description": "Portée : read — Consultation. Tout jeton (sauf backup/sync seuls) et tout rôle.",
        "operationId": "get_api_groups",
        "x-rempart-scope": "read",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "Groupes d'appareils"
        ],
        "summary": "Créer un groupe",
        "description": "Portée : admin — Modification de la configuration. Jeton « admin », rôle opérateur ou administrateur.",
        "operationId": "post_api_groups",
        "x-rempart-scope": "admin",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "clients": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "blocking": {
                    "type": "boolean"
                  },
                  "inherit_lists": {
                    "type": "boolean"
                  },
                  "lists": {
                    "type": "array",
                    "items": {}
                  },
                  "inherit_rules": {
                    "type": "boolean"
                  },
                  "rules": {
                    "type": "array",
                    "items": {}
                  },
                  "services": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "schedules": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "days": {
                          "type": "array",
                          "items": {
                            "type": "integer"
                          }
                        },
                        "start": {
                          "type": "string"
                        },
                        "end": {
                          "type": "string"
                        },
                        "services": {
                          "type": "array",
                          "items": {}
                        },
                        "block_all": {
                          "type": "boolean"
                        }
                      }
                    }
                  },
                  "safe_search": {
                    "type": "boolean"
                  },
                  "youtube": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "name": "Enfants",
                "clients": [
                  "192.168.1.50",
                  "aa:bb:cc:dd:ee:ff",
                  "192.168.10.0/24"
                ],
                "blocking": true,
                "inherit_lists": true,
                "lists": [],
                "inherit_rules": true,
                "rules": [],
                "services": [
                  "tiktok"
                ],
                "schedules": [
                  {
                    "name": "Nuit",
                    "days": [
                      0,
                      1,
                      2,
                      3,
                      4,
                      5,
                      6
                    ],
                    "start": "22:00",
                    "end": "07:00",
                    "services": [],
                    "block_all": true
                  }
                ],
                "safe_search": true,
                "youtube": "moderate"
              }
            }
          }
        }
      }
    },
    "/api/groups/{id}": {
      "put": {
        "tags": [
          "Groupes d'appareils"
        ],
        "summary": "Remplacer un groupe",
        "description": "Portée : admin — Modification de la configuration. Jeton « admin », rôle opérateur ou administrateur.",
        "operationId": "put_api_groups_id",
        "x-rempart-scope": "admin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "clients": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "blocking": {
                    "type": "boolean"
                  },
                  "inherit_lists": {
                    "type": "boolean"
                  },
                  "lists": {
                    "type": "array",
                    "items": {}
                  },
                  "inherit_rules": {
                    "type": "boolean"
                  },
                  "rules": {
                    "type": "array",
                    "items": {}
                  },
                  "services": {
                    "type": "array",
                    "items": {}
                  },
                  "schedules": {
                    "type": "array",
                    "items": {}
                  },
                  "safe_search": {
                    "type": "boolean"
                  },
                  "youtube": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "id": "…",
                "name": "Enfants",
                "clients": [
                  "192.168.1.50"
                ],
                "blocking": true,
                "inherit_lists": true,
                "lists": [],
                "inherit_rules": true,
                "rules": [],
                "services": [],
                "schedules": [],
                "safe_search": true,
                "youtube": ""
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Groupes d'appareils"
        ],
        "summary": "Supprimer un groupe",
        "description": "Portée : admin — Modification de la configuration. Jeton « admin », rôle opérateur ou administrateur.",
        "operationId": "delete_api_groups_id",
        "x-rempart-scope": "admin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Jeton rmp_… créé dans Sécurité → Accès API ou par POST /api/tokens."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rempart_session",
        "description": "Cookie posé par POST /api/login."
      }
    },
    "responses": {
      "Error": {
        "description": "Erreur",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}